+++ PolyToEigen/LinearLS.h
@@ -14,7 +14,7 @@ namespace Triangulation {
 		 *	\param	p1	Corresponding point in the image of the second camera.
 		 *	\return	Triangulated point.
 		 */
-		Eigen::Vector3d triangulate(const Eigen::Vector2d& p0, const Eigen::Vector2d& p1) const override;
+		Eigen::Vector3f triangulate(const Eigen::Vector2f& p0, const Eigen::Vector2f& p1) const override;
 	};
 #endif 
 }
