+++ PolyToEigen/PolyBase.h
@@ -7,14 +7,14 @@
 #include <vector>
 
 
-using Eigen::Matrix3d;
-using Eigen::Vector3d;
+using Eigen::Matrix3f;
+using Eigen::Vector3f;
 
 namespace Triangulation {
 	class PolyBase : public TriangulationBase
 	{
-		typedef Eigen::Matrix3d Intrinsic;
-		typedef Eigen::Matrix3d Fundamental;
+		typedef Eigen::Matrix3f Intrinsic;
+		typedef Eigen::Matrix3f Fundamental;
 
 		/**
 		 *	\brief	Constructor
@@ -24,7 +24,7 @@ namespace Triangulation {
 		
 
 		public:
-		PolyBase(const Eigen::MatrixXd& P0, const Eigen::MatrixXd& P1);
+		PolyBase(const Eigen::MatrixXf& P0, const Eigen::MatrixXf& P1);
 			/**
 			 *	\brief	Constructor
 			 *	\param	P0	Camera matrix of the first camera.
@@ -39,56 +39,56 @@ namespace Triangulation {
 		 *	\param 	P1	Camera matrix of the second camera.
 		 */
 		
-		PolyBase(const Eigen::MatrixXd& P0, const Eigen::MatrixXd& P1, const Fundamental& F);
+		PolyBase(const Eigen::MatrixXf& P0, const Eigen::MatrixXf& P1, const Fundamental& F);
 			/**
 			 *	\brief	Triangulates image points.
 			 *	\param	p0	Point in the image of the first camera.
 			 *	\param	p1	Corresponding point in the image of the second camera.
 			 *	\return	Triangulated point.
 			 */
-		Eigen::Vector3d triangulate(const Eigen::Vector2d& p0, const Eigen::Vector2d& p1) const override;
+		Eigen::Vector3f triangulate(const Eigen::Vector2f& p0, const Eigen::Vector2f& p1) const override;
 	protected:
-		typedef Eigen::Vector3d Line;
-		typedef Eigen::Vector3d Epipole;
-		typedef std::vector<std::complex<double> > Roots;
-		typedef std::tuple<double, double, double, double, double, double> PolyParams;
+		typedef Eigen::Vector3f Line;
+		typedef Eigen::Vector3f Epipole;
+		typedef std::vector<std::complex<float> > Roots;
+		typedef std::tuple<float, float, float, float, float, float> PolyParams;
 		/**
 		 *	\brief	Computes corrected correspondences, that minimize the geometric error.
 		 *	\param	p0	Point in the image of the first camera.
 		 *	\param	p1	Corresponding point in the image of the second camera.
 		 *	\return	Corrected correspondences in homogeneous coordinates.
 		 */
-		std::pair<Eigen::Vector2d, Eigen::Vector2d> ComputeCorrectedCorrespondences(const Eigen::Vector2d& p0, const Eigen::Vector2d& p1) const;
+		std::pair<Eigen::Vector2f, Eigen::Vector2f> ComputeCorrectedCorrespondences(const Eigen::Vector2f& p0, const Eigen::Vector2f& p1) const;
 		/**
 		 *	\brief	Defines translation matrix that translate the given point to origin.
 		 *	\param	p	Translated point.
 		 *	\return	Translation matrix.
 		 */
-		Eigen::MatrixXd TranslateToOrigin(const Eigen::Vector2d& p) const;
+		Eigen::MatrixXf TranslateToOrigin(const Eigen::Vector2f& p) const;
 		/**
 		 *	\brief	Computes epipole e = (e1, e2, e3) such as eF = 0 and e1*e1 + e2*e2 = 1
 		 *	\param	F	Fundamental matrix.
 		 *	\return	Computed epipole
 		 */
-		Epipole ComputeLeftEpipole(const Eigen::MatrixXd& F) const;
+		Epipole ComputeLeftEpipole(const Eigen::MatrixXf& F) const;
 		/**
 		 *	\brief	Computes epipole e = (e1, e2, e3) such as Fe = 0 and e1*e1 + e2*e2 = 1
 		 *	\param	F	Fundamental matrix.
 		 *	\return	Computed epipole
 		 */
-		Epipole ComputeRightEpipole(const Eigen::MatrixXd& F) const;
+		Epipole ComputeRightEpipole(const Eigen::MatrixXf& F) const;
 		/**
 		 *	\brief	Defines rotation matrix using given epipole.
 		 *	\param	e	Epipole.
 		 *	\return	Rotation matrix.
 		 */
-		Eigen::MatrixXd FormRotationMatrix(const Epipole& e) const;
+		Eigen::MatrixXf FormRotationMatrix(const Epipole& e) const;
 		/**
 		 *	\brief	Prepares polynomial coefficients.
 		 *	\param	params	Polynomial coefficients params.
 		 *	\return Polynomial coefficients.
 		 */
-		virtual std::vector<double> PreparePolyCoeffs(const PolyParams& params) const = 0;
+		virtual std::vector<float> PreparePolyCoeffs(const PolyParams& params) const = 0;
 		/**
 		 *	\brief	Forms and solves 6th degree polynomial.
 		 *	\param	params	Polynomial coefficients params.
@@ -101,27 +101,27 @@ namespace Triangulation {
 		 *	\param	params	Polynomial coefficients params.
 		 *	\return	Array of cost for each root.
 		 */
-		virtual std::vector<double> EvaluateRootsCosts(const Roots& roots, const PolyParams& params) const = 0;
+		virtual std::vector<float> EvaluateRootsCosts(const Roots& roots, const PolyParams& params) const = 0;
 		/**
 		 *	\brief	Evaluates cost function for roots.
 		 *	\param	roots	Six roots of 6th degree polynomial.
 		 *	\param	params	Polynomial coefficients params.
 		 *	\return	Root that gives smallest cost function value.
 		 */
-		double EvaluateRoots(const Roots& roots, const PolyParams& params) const;
+		float EvaluateRoots(const Roots& roots, const PolyParams& params) const;
 		/**
 		 *	\brief	Construct two epipolar lines on which the corrected correspondences lie.
 		 *	\param	t	Real root for which the cost function gives the smallest value.
 		 *	\param	params	Polynomial coefficients params.
 		 *	\return	Pair of epipolar lines.
 		 */
-		std::pair<Line, Line> ConstructLines(double t, const PolyParams& params) const;
+		std::pair<Line, Line> ConstructLines(float t, const PolyParams& params) const;
 		/**
 		 *	\brief	Finds point on given line that is closest to the origin.
 		 *	\param	l	Line on which the point lies.
 		 *	\return	Point on line closest to the origin.
 		 */
-		Eigen::Vector3d FindPointOnLineClosestToOrigin(const Line& l) const;
+		Eigen::Vector3f FindPointOnLineClosestToOrigin(const Line& l) const;
 		/**
 		 *	\brief	Transfers point to original coordinates using R and T.
 		 *	\param	p	A point to transfer.
@@ -129,7 +129,7 @@ namespace Triangulation {
 		 *	\param	T	Translational transformation.
 		 *	\return 	Point in original coordinates.
 		 */
-		Eigen::Vector3d TransferPointToOriginalCoordinates(const Eigen::Vector3d& p, const Eigen::MatrixXd& R, const Eigen::MatrixXd& T) const;
+		Eigen::Vector3f TransferPointToOriginalCoordinates(const Eigen::Vector3f& p, const Eigen::MatrixXf& R, const Eigen::MatrixXf& T) const;
 		/**
 		 *	\brief	Sets origin of world coordinate system to first camera.
 		 *	\param	P0	First camera projection matrix.
@@ -137,25 +137,25 @@ namespace Triangulation {
 		 *	\return	K0, K1, R and T - camera intrinsics and orientation of second camera in new world coordinates.
 		 */
 
-		std::tuple<Intrinsic, Intrinsic, Eigen::MatrixXd, Eigen::MatrixXd> SetOriginToCamera(const Eigen::MatrixXd& P0, const Eigen::MatrixXd& P1) const;		/**
+		std::tuple<Intrinsic, Intrinsic, Eigen::MatrixXf, Eigen::MatrixXf> SetOriginToCamera(const Eigen::MatrixXf& P0, const Eigen::MatrixXf& P1) const;		/**
 		 *	\brief	Computes fundamental matrix from camera projection matrices.
 		 *	\param	P0	First camera projection matrix.
 		 *	\param	P1	Second camera projection matrix.
 		 *	\return	Computed fundamental matrix.
 		 */
-		Fundamental ComputeFundamentalMatrix(const Eigen::MatrixXd& P0, const Eigen::MatrixXd& P1) const;
+		Fundamental ComputeFundamentalMatrix(const Eigen::MatrixXf& P0, const Eigen::MatrixXf& P1) const;
 		/**
 		 *	\brief	Returns the order of the polynomial with given coefficients (highest non-zero coeffs index).
 		 *	\param	coeffs	Polynomial coefficients.
 		 *	\return	Polynomial order
 		 */
-		int FindPolynomialOrder(const std::vector<double>& coeffs) const;
+		int FindPolynomialOrder(const std::vector<float>& coeffs) const;
 		/**
 		 *	\brief	Returns canonic camera projection matrix of second camera computed from given fundamental matrix.
 		 *	\param	F	Fundamental matrix.
 		 *	\return	Canonic	camera projection matrix.
 		 */
-		Eigen::MatrixXd CameraProjectionMatrixFromFundamentalMatrix(const Fundamental& F) const;
+		Eigen::MatrixXf CameraProjectionMatrixFromFundamentalMatrix(const Fundamental& F) const;
 
 
 		const Fundamental F;
